requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.3.0
APScheduler==3.10.4
astunparse==1.6.3
attrs==24.2.0
Brotli==1.1.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.0
curl_cffi==0.7.2
cycler==0.12.1
deprecation==2.1.0
fastapi==0.110.0
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.4.1
g4f==0.3.2.9
gast==0.6.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.66.1
h11==0.14.0
h5py==3.12.1
httptools==0.6.1
idna==3.8
joblib==1.4.2
keras==3.6.0
kiwisolver==1.4.7
libclang==18.1.1
lxml==5.3.0
magic-filter==1.0.12
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.3.2
mplfinance==0.12.10b0
multidict==6.1.0
namex==0.0.8
numpy==1.26.4
opt_einsum==3.4.0
optree==0.13.0
packaging==24.1
pandas==2.2.2
pandas-datareader==0.10.0
pillow==10.4.0
propcache==0.2.0
protobuf==4.25.4
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyparsing==3.1.4
pyTelegramBotAPI==4.23.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
rich==13.9.2
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.36.3
ta==0.11.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
termcolor==2.4.0
threadpoolctl==3.5.0
tinkoff==0.1.1
tinkoff-investments==0.2.0b105
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
uvicorn==0.29.0
# uvloop==0.19.0  # Not supported on Windows
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.4
wheel==0.44.0
wrapt==1.16.0
yarl==1.11.1
